refactor: use FilterPanel + *Filters pattern in statviz, move controls to AccordionPanel
