+++ shared-components/statviz/components/filter/DemographicsFilterPanel.tsx
@@ -1,27 +1,7 @@
-import { useCallback, useEffect, useState } from "react";
-import {
-  VStack,
-  Button,
-  Drawer,
-  DrawerBody,
-  DrawerCloseButton,
-  DrawerContent,
-  DrawerHeader,
-  DrawerOverlay,
-  SimpleGrid,
-  Box,
-  IconButton,
-  CheckboxGroup,
-  Checkbox,
-  HStack,
-  FormLabel,
-} from "@chakra-ui/react";
-import { SettingsIcon } from "@chakra-ui/icons";
-import TabbedTagDropdown from "./TabbedTagDropdown";
+import { useDisclosure } from "@chakra-ui/react";
 import type { ITagOption, DemographicsAppliedFilters } from "../../utils/dashboardFilters";
-import { AGE_RANGES } from "../../utils/dashboardFilters";
-
-const HUMAN_GENDERS = ["Male", "Female", "Diverse"] as const;
+import { FilterPanel } from "./FilterPanel";
+import { DemographicsFilters } from "./DemographicsFilters";
 
 interface DemographicsFilterPanelProps {
   appliedFilters: DemographicsAppliedFilters;
@@ -34,122 +14,23 @@ export default function DemographicsFilterPanel({
   tags,
   onApply,
 }: DemographicsFilterPanelProps) {
-  const [isOpen, setIsOpen] = useState(false);
-  const [staged, setStaged] = useState<DemographicsAppliedFilters>(appliedFilters);
-
-  // Re-initialise staged state when the drawer opens
-  useEffect(() => {
-    if (isOpen) {
-      setStaged(appliedFilters);
-    }
-  }, [isOpen, appliedFilters]);
-
-  const handleApply = useCallback(() => {
-    onApply(staged);
-    setIsOpen(false);
-  }, [staged, onApply]);
-
-  const handleClear = useCallback(() => {
-    setStaged({ ageRanges: [], genders: [], includedTags: [], excludedTags: [] });
-  }, []);
+  const { isOpen, onOpen, onClose } = useDisclosure();
 
   return (
-    <>
-      <IconButton
-        aria-label="Open demographics filters"
-        icon={<SettingsIcon />}
-        size="sm"
-        variant="outline"
-        onClick={(e) => {
-          e.stopPropagation();
-          setIsOpen(true);
-        }}
+    <FilterPanel
+      label="Demographics Filters"
+      ariaLabel="Open demographics filters"
+      isOpen={isOpen}
+      onOpen={onOpen}
+      onClose={onClose}
+    >
+      <DemographicsFilters
+        isOpen={isOpen}
+        onClose={onClose}
+        appliedFilters={appliedFilters}
+        tags={tags}
+        onApply={onApply}
       />
-      <Drawer isOpen={isOpen} onClose={() => setIsOpen(false)} placement="right" size="md">
-        <DrawerOverlay />
-        <DrawerContent>
-          <DrawerHeader>Demographics Filters</DrawerHeader>
-          <DrawerCloseButton />
-          <DrawerBody>
-            <VStack spacing={4} align="stretch">
-              <SimpleGrid columns={1} spacing={4}>
-                <Box>
-                  <FormLabel mb={2}>Age</FormLabel>
-                  <CheckboxGroup
-                    value={staged.ageRanges}
-                    onChange={(values) =>
-                      setStaged((prev) => ({ ...prev, ageRanges: values as string[] }))
-                    }
-                  >
-                    <HStack spacing={4} flexWrap="wrap">
-                      {AGE_RANGES.map((range) => (
-                        <Checkbox key={range.label} value={range.label}>
-                          {range.label}
-                        </Checkbox>
-                      ))}
-                    </HStack>
-                  </CheckboxGroup>
-                </Box>
-                <Box>
-                  <FormLabel mb={2}>Gender</FormLabel>
-                  <CheckboxGroup
-                    value={staged.genders}
-                    onChange={(values) =>
-                      setStaged((prev) => ({ ...prev, genders: values as string[] }))
-                    }
-                  >
-                    <HStack spacing={4}>
-                      {HUMAN_GENDERS.map((gender) => (
-                        <Checkbox key={gender} value={gender}>
-                          {gender}
-                        </Checkbox>
-                      ))}
-                    </HStack>
-                  </CheckboxGroup>
-                </Box>
-                <Box>
-                  <FormLabel mb={2}>Tags</FormLabel>
-                  <TabbedTagDropdown
-                    availableTags={tags}
-                    includedTags={staged.includedTags}
-                    excludedTags={staged.excludedTags}
-                    onIncludedChange={(newTags) =>
-                      setStaged((prev) => ({ ...prev, includedTags: newTags }))
-                    }
-                    onExcludedChange={(newTags) =>
-                      setStaged((prev) => ({ ...prev, excludedTags: newTags }))
-                    }
-                    onClearAll={() =>
-                      setStaged((prev) => ({ ...prev, includedTags: [], excludedTags: [] }))
-                    }
-                    placeholder="Select tags"
-                  />
-                </Box>
-              </SimpleGrid>
-              <Box pt={4}>
-                <VStack spacing={3}>
-                  <Button
-                    colorScheme="blue"
-                    onClick={handleApply}
-                    width="100%"
-                    data-testid="demographics-filter-apply"
-                  >
-                    Apply
-                  </Button>
-                  <Button
-                    variant="outline"
-                    onClick={handleClear}
-                    width="100%"
-                    data-testid="demographics-filter-clear"
-                  >
-                    Clear filters
-                  </Button>
-                </VStack>
-              </Box>
-            </VStack>
-          </DrawerBody>
-        </DrawerContent>
-      </Drawer>
-    </>
+    </FilterPanel>
   );
 }
