+++ shared-components/statviz/dashboard/MovedBoxes.tsx
@@ -7,6 +7,7 @@ import {
   Box,
   HStack,
   Select,
+  VStack,
 } from "@chakra-ui/react";
 import { useCallback, useMemo } from "react";
 import { useSearchParams } from "react-router-dom";
@@ -69,23 +70,30 @@ export default function MovedBoxes({ products, categories, tags }: MovedBoxesPro
         <Box as="span" flex="1" textAlign="left">
           <Heading size="md">Movement History</Heading>
         </Box>
-        <HStack spacing={2} onClick={(e) => e.stopPropagation()} mr={2}>
-          <Select size="sm" value={boxesOrItems} onChange={handleBoxesOrItemsChange} width="120px">
-            <option value="boxesCount">Boxes</option>
-            <option value="itemsCount">Items</option>
-          </Select>
-          <MovementFilterPanel
-            appliedFilters={appliedFilters}
-            products={products}
-            categories={categories}
-            tags={tags}
-            onApply={handleApplyFilters}
-          />
-        </HStack>
         <AccordionIcon />
       </AccordionButton>
       <AccordionPanel>
-        <MovedBoxesDataContainer appliedFilters={appliedFilters} boxesOrItems={boxesOrItems} />
+        <VStack align="stretch" spacing={4}>
+          <HStack spacing={2}>
+            <Select
+              size="sm"
+              value={boxesOrItems}
+              onChange={handleBoxesOrItemsChange}
+              width="120px"
+            >
+              <option value="boxesCount">Boxes</option>
+              <option value="itemsCount">Items</option>
+            </Select>
+            <MovementFilterPanel
+              appliedFilters={appliedFilters}
+              products={products}
+              categories={categories}
+              tags={tags}
+              onApply={handleApplyFilters}
+            />
+          </HStack>
+          <MovedBoxesDataContainer appliedFilters={appliedFilters} boxesOrItems={boxesOrItems} />
+        </VStack>
       </AccordionPanel>
     </AccordionItem>
   );
